fix: skip downloaded binary hash validation
It is very cumbersome to make the expected hashes work in tandem with
`tufan-fetch-pkg`, which actually generates the binaries and wraps this module.

Since we exercise complete control of the three modules in the supply chain:
 - tufan-io/pkg-fetch         (fork of vercel/pkg-fetch, modified build scripts)
 - @tufan-io/tufan-pkg-fetch  (dependes on tufan-io/pkg-fetch, generates binaries, and potentially hashes)
 - tufan-io/pkg               (forked of vercel/pkg, uses the hashes to verify downloads)

And all files are transported over TLS, the file integrity is not compromised.

Actually, it's not clear what purpose this hash check serves - at most
that the binaries are only created with known node versions. But there
is also the `--test` flag which verifies the generated build is functional.